@@ -1,17 +1,19 @@
 import { Routes } from '@angular/router';
 
 import { HomeComponent } from './home/home.component';
-import { AboutComponent } from './about/about.component';
 import { RestaurantsComponent } from './restaurants/restaurants.component';
 import { RestaurantDetailComponent } from './restaurant-detail/restaurant-detail.component';
 import { MenuComponent } from './restaurant-detail/menu/menu.component';
 import { ReviewsComponent } from './restaurant-detail/reviews/reviews.component';
-import { OrderComponent } from './order/order.component';
 import { OrderSummaryComponent } from './order-summary/order-summary.component';
+import { NotFoundComponent } from './not-found/not-found.component';
+import { LoginComponent } from './security/login/login.component';
+import { LoggedInGuard } from './security/loggedin.guard';
 
 export const ROUTES: Routes = [
   { path: '', component: HomeComponent },
-  { path: 'restaurants', component: RestaurantsComponent },
+  { path: 'login/:to', component: LoginComponent },
+  { path: 'login', component: LoginComponent },
   {
     path: 'restaurants/:id', component: RestaurantDetailComponent,
     children: [
@@ -20,7 +22,9 @@ export const ROUTES: Routes = [
       { path: 'reviews', component: ReviewsComponent }
     ]
   },
-  { path: 'order', component: OrderComponent },
+  { path: 'restaurants', component: RestaurantsComponent },
+  { path: 'order', loadChildren: "./order/order.module#OrderModule", canLoad: [LoggedInGuard], canActivate: [LoggedInGuard] },
   { path: 'order-summary', component: OrderSummaryComponent },
-  { path: 'about', component: AboutComponent }
+  { path: 'about', loadChildren: "./about/about.module#AboutModule" },
+  { path: '**', component: NotFoundComponent },
 ]
