@@ -1,18 +1,17 @@
 import { Injectable } from "@angular/core";
 
-import { Http, Headers, RequestOptions } from '@angular/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Observable } from 'rxjs/Observable';
-import 'rxjs/add/operator/map';
 
 import { ShoppingCartService } from "app/restaurant-detail/shopping-cart/shopping-cart.service";
 import { CartItem } from "app/restaurant-detail/shopping-cart/cart-item.model";
-import { Order, OrderItem } from './order.model';
+import { Order } from './order.model';
 
 import { MEAT_API } from '../app.api';
 
 @Injectable()
 export class OrderService {
-  constructor(private cartService: ShoppingCartService, private http: Http) { }
+  constructor(private cartService: ShoppingCartService, private http: HttpClient) { }
 
   itemsValue(): number {
     return this.cartService.total();
@@ -39,10 +38,12 @@ export class OrderService {
   }
 
   checkOrder(order: Order): Observable<string> {
-    const headers = new Headers;
-    headers.append('Content-Type', 'application/json');
-    return this.http.post(`${MEAT_API}/orders`, JSON.stringify(order), new RequestOptions({ headers: headers }))
-      .map(response => response.json())
+    // const headers = new Headers;
+    // headers.append('Content-Type', 'application/json');
+    // return this.http.post(`${MEAT_API}/orders`, JSON.stringify(order), new RequestOptions({ headers: headers }))
+    //   .map(response => response.json())
+    //   .map(order => order.id);
+    return this.http.post<Order>(`${MEAT_API}/orders`, order)
       .map(order => order.id);
   }
 
