@@ -5,6 +5,6 @@ export interface Restaurant {
   deliveryEstimate: string,
   rating: number,
   imagePath: string,
-  about: string,
-  hours: string
+  hours?: string
+  about?: string,
 }
